+++ src/assets/ba_data/python/baenv.py
@@ -50,7 +50,7 @@ if TYPE_CHECKING:
 
 # Build number and version of the ballistica binary we expect to be
 # using.
-TARGET_BALLISTICA_BUILD = 21204
+TARGET_BALLISTICA_BUILD = 21206
 TARGET_BALLISTICA_VERSION = '1.7.25'