+++ src/ballistica/shared/ballistica.cc
@@ -39,7 +39,7 @@ auto main(int argc, char** argv) -> int {
 namespace ballistica {
 
 // These are set automatically via script; don't modify them here.
-const int kEngineBuildNumber = 21204;
+const int kEngineBuildNumber = 21206;
 const char* kEngineVersion = "1.7.25";
 
 #if BA_MONOLITHIC_BUILD