+++ tools/batools/docs.py
@@ -2,6 +2,8 @@
 #
 """Documentation generation functionality."""
 
+# pyright: reportPrivateImportUsage=false
+
 from __future__ import annotations
 
 import os